Chinese book cover adjunct
This is the book cover I knocked up for a book I partially translated in to Chinese. As a creative aside. Also I once fell in love with a Chinese woman. And got Mandarin/Shaolin scrolls up the sides of my arm. Isn't it funny, the things we do for love?!